% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Compute a correlation matrix with bootstrap confidence limits}
\usage{
correlation(
  x,
  method = "spearman",
  summary = "actual",
  alpha = 0.05,
  B = 1000,
  corfun = cor,
  digits = 3
)
}
\arguments{
\item{x}{A numeric matrix or data frame whose correlation matrix is required.}

\item{method}{Argument to the function that computes the correlation. Defaults
to \code{method="spearman"}.}

\item{alpha}{Determines the size of the confidence interval. Defaults to
\code{alpha=0.05} and a 95\% interval is computed.}

\item{B}{The number of bootstrap samples to run. Defaults to \code{B=1000}.}

\item{corfun}{The function to be used to compute the correlation. Defaults to
\code{corfun=cor}. If you substitute in another (e.g. a robust correlation
function) it must have an argument called 'method' so you might need to
write a wrapper for it.}
}
\description{
Compute a correlation matrix with bootstrap confidence limits
}
