% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatTable}
\alias{formatTable}
\title{Format a matrix or data frame}
\usage{
formatTable(x, types = NULL, fmt = "\%.4f", NAstring = "")
}
\arguments{
\item{x}{A matrix or data frame.}

\item{types}{Character vector having the same number of elements as there are
columns in \code{x}.}

\item{fmt}{To be passed to \code{sprintf}, defaults to \code{fmt = "\%.4f"}.}

\item{NAstring}{String to replace \code{NA}s. Defaults to \code{NAstring = ""}.}
}
\value{
A data frame, the columns having been formatted accordingly
}
\description{
Format a matrix or data frame
}
\details{
The allowed types are "i" (integer), "n" (numeric), "p" (p-values),
  and "c" (character). If the type is "i" a character representation of the
  values coerced to integer is returned; if type is "n" a character representation
  of the values after passing through \code{sprintf(fmt, x)} is retunred;
  if type is "p", values less than 0.0001 are described as such (i.e. "<0.0001"),
  otherwise the returned values from \code{sptritf(fmt, X)}; if type is "c",
  the values are returned unaltered.
}
