% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{getPstring}
\alias{getPstring}
\title{Get a string to represent small p-values}
\usage{
getPstring(f = "\%.4f")
}
\arguments{
\item{f}{A string of the type passed into \code{sprintf}.}
}
\description{
Get a string to represent small p-values
}
\details{
This function expects its input to be of the form "xxnnnx" where the
  xs are stripped off and teh ns are coerced to numeric.
}
