% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTable.R
\name{formatModelTable}
\alias{formatModelTable}
\title{Format a summary table from a model}
\usage{
formatModelTable(x, fmt = "\%.4f")
}
\arguments{
\item{x}{An object summarizing a model, either a matrix or data frame.}
}
\description{
Format a summary table from a model
}
\details{
It is assumed that the 4th column is p-values, and the function
  either formats them using \code{sprintf} (as it does to other columns),
  or turns them into "<0.0001".
}
